Created separate database callback for separate releases
