+++ core/src/main/java/org/kiwix/kiwixmobile/core/data/KiwixRoomDatabase.kt
@@ -19,62 +19,39 @@
 package org.kiwix.kiwixmobile.core.data
 
 import android.content.Context
-import android.util.Log
 import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
-import androidx.sqlite.db.SupportSQLiteDatabase
-import io.objectbox.BoxStore
-import io.objectbox.kotlin.boxFor
 import org.kiwix.kiwixmobile.core.BuildConfig
 import org.kiwix.kiwixmobile.core.dao.RecentSearchRoomDao
 import org.kiwix.kiwixmobile.core.dao.entities.RecentSearchRoomEntity
+import org.kiwix.kiwixmobile.core.data.remote.FdroidDatabaseCallback
+import org.kiwix.kiwixmobile.core.data.remote.RoomDatabaseCallback
 
 @Suppress("UnnecessaryAbstractClass")
 @Database(entities = [RecentSearchRoomEntity::class], version = 1)
 abstract class KiwixRoomDatabase : RoomDatabase() {
   abstract fun recentSearchRoomDao(): RecentSearchRoomDao
-  abstract fun foo(): KiwixRoomDatabase
 
   companion object {
     private var db: KiwixRoomDatabase? = null
-    abstract fun foo()
-    fun getInstance(context: Context, boxStore: BoxStore): KiwixRoomDatabase {
+    fun getInstance(context: Context): KiwixRoomDatabase {
+      val callback =
+        if (BuildConfig.BUILD_TYPE != "fdroid")
+          RoomDatabaseCallback(context)
+        else FdroidDatabaseCallback()
       return db ?: synchronized(KiwixRoomDatabase::class) {
         return@getInstance db
           ?: Room.databaseBuilder(context, KiwixRoomDatabase::class.java, "KiwixRoom.db")
             // We have already database name called kiwix.db in order to avoid complexity we named as
             // kiwixRoom.db
-            .addCallback(object : RoomDatabase.Callback() {
-              override fun onCreate(db: SupportSQLiteDatabase) {
-                super.onCreate(db)
-                Log.d("gouri", "onCreate")
-              }
-
-              override fun onOpen(db: SupportSQLiteDatabase) {
-                super.onOpen(db)
-                Log.d("gouri", "onOpen")
-              }
-
-              override fun onDestructiveMigration(db: SupportSQLiteDatabase) {
-                super.onDestructiveMigration(db)
-                Log.d("gouri", "onDestructiveMigration")
-              }
-            })
-            .build().also {
-              if (!BuildConfig.BUILD_TYPE.contentEquals("fdroid")) {
-                it.migrateRecentSearch(boxStore)
-              }
-            }
+            .addCallback(callback)
+            .build()
       }
     }
 
     fun destroyInstance() {
       db = null
     }
-  }
-
-  fun migrateRecentSearch(boxStore: BoxStore) {
-    recentSearchRoomDao().migrationToRoomInsert(boxStore.boxFor())
   }
 }
