+++ core/src/main/java/org/kiwix/kiwixmobile/core/di/components/CoreComponent.kt
@@ -37,6 +37,7 @@ import org.kiwix.kiwixmobile.core.dao.RecentSearchRoomDao
 import org.kiwix.kiwixmobile.core.data.DataModule
 import org.kiwix.kiwixmobile.core.data.DataSource
 import org.kiwix.kiwixmobile.core.data.remote.KiwixService
+import org.kiwix.kiwixmobile.core.data.remote.RoomDatabaseCallback
 import org.kiwix.kiwixmobile.core.di.modules.ApplicationModule
 import org.kiwix.kiwixmobile.core.di.modules.CoreViewModelModule
 import org.kiwix.kiwixmobile.core.di.modules.JNIModule
@@ -105,6 +106,7 @@ interface CoreComponent {
 
   fun inject(errorActivity: ErrorActivity)
   fun inject(searchFragment: SearchFragment)
+  fun inject(roomDatabaseCallback: RoomDatabaseCallback)
 
   fun inject(settingsFragment: CoreSettingsFragment)
   fun coreServiceComponent(): CoreServiceComponent.Builder
